requirements.txt
# Bankruptcy Query Optimizer Requirements

# Core dependency
openai-agents>=0.2.0

# Async support (if not already installed)
asyncio>=3.4.3

# Data validation
pydantic>=2.0.0

# Web search and content extraction
brave-search>=0.2.0
httpx>=0.25.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
python-dotenv>=1.0.0

# Optional dependencies for development
pytest>=7.0.0
pytest-asyncio>=0.21.0